# app/static/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api/ping": {
            "get": {
                "summary": "Ping",
                "operationId": "ping_api_ping_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Ping Api Ping Get",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ping-with-token": {
            "get": {
                "summary": "Ping With Token",
                "operationId": "ping_with_token_api_ping_with_token_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Ping With Token Api Ping With Token Get",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "summary": "Signin",
                "operationId": "signin_auth_signin_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Signin Auth Signin Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "delete": {
                "summary": "Logout",
                "operationId": "logout_auth_logout_delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Logout Auth Logout Delete",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/is_auth": {
            "get": {
                "summary": "Is Auth",
                "operationId": "is_auth_auth_is_auth_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Is Auth Auth Is Auth Get",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/resecure": {
            "post": {
                "summary": "Resecure",
                "operationId": "resecure_auth_resecure_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Resecure Auth Resecure Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/OKResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parse/data": {
            "post": {
                "summary": "Parse Html",
                "operationId": "parse_html_parse_data_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParsingRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Parse Html Parse Data Post",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ParsingResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parse/history": {
            "get": {
                "summary": "Parse Html",
                "operationId": "parse_html_parse_history_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Parse Html Parse History Get",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ParsingHistoryResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parse/history/{id}": {
            "get": {
                "summary": "Parse Html",
                "operationId": "parse_html_parse_history__id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Id",
                            "type": "integer"
                        },
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Parse Html Parse History  Id  Get",
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/ParsingResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorCode": {
                "title": "ErrorCode",
                "enum": [
                    -1,
                    1,
                    2,
                    3
                ],
                "type": "integer",
                "description": "An enumeration."
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "ERROR"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string",
                        "default": ""
                    },
                    "error_code": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ErrorCode"
                            }
                        ],
                        "default": -1
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "OKResponse": {
                "title": "OKResponse",
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "OK"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string",
                        "default": ""
                    }
                }
            },
            "ParsingElement": {
                "title": "ParsingElement",
                "required": [
                    "word",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "word": {
                        "title": "Word",
                        "type": "string"
                    },
                    "value": {
                        "title": "Value",
                        "type": "integer"
                    }
                }
            },
            "ParsingElementData": {
                "title": "ParsingElementData",
                "required": [
                    "elements"
                ],
                "type": "object",
                "properties": {
                    "elements": {
                        "title": "Elements",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ParsingElement"
                        }
                    }
                }
            },
            "ParsingHistory": {
                "title": "ParsingHistory",
                "required": [
                    "id",
                    "url"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "url": {
                        "title": "Url",
                        "type": "string"
                    }
                }
            },
            "ParsingHistoryResponse": {
                "title": "ParsingHistoryResponse",
                "required": [
                    "data"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "OK"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string",
                        "default": ""
                    },
                    "data": {
                        "title": "Data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ParsingHistory"
                        }
                    }
                }
            },
            "ParsingRequest": {
                "title": "ParsingRequest",
                "required": [
                    "url"
                ],
                "type": "object",
                "properties": {
                    "url": {
                        "title": "Url",
                        "type": "string"
                    },
                    "save": {
                        "title": "Save",
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "ParsingResponse": {
                "title": "ParsingResponse",
                "required": [
                    "url",
                    "data"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Status"
                            }
                        ],
                        "default": "OK"
                    },
                    "message": {
                        "title": "Message",
                        "type": "string",
                        "default": ""
                    },
                    "url": {
                        "title": "Url",
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ParsingElementData"
                    }
                }
            },
            "Status": {
                "title": "Status",
                "enum": [
                    "OK",
                    "ERROR"
                ],
                "type": "string",
                "description": "An enumeration."
            },
            "UserRequest": {
                "title": "UserRequest",
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}